Amelioration auth

+++ src/AppBundle/Controller/TaskController.php
@@ -3,6 +3,7 @@
 namespace AppBundle\Controller;
 
 use AppBundle\Entity\Task;
+use AppBundle\Entity\User;
 use AppBundle\Form\TaskType;
 use AppBundle\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -16,6 +17,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Twig\Environment;
 
 class TaskController extends Controller
@@ -36,6 +38,9 @@ class TaskController extends Controller
     /** @var UrlGeneratorInterface */
     private $urlGenerator;
 
+    /** @var AuthorizationCheckerInterface */
+    private $authorizationChecker;
+
     /** @var FlashBagInterface */
     private $flashBag;
 
@@ -49,27 +54,35 @@ class TaskController extends Controller
      * @param FormFactoryInterface $formFactory
      * @param EntityManagerInterface $entityManager
      * @param UrlGeneratorInterface $urlGenerator
+     * @param AuthorizationCheckerInterface $authorizationChecker
      * @param FlashBagInterface $flashBag
      * @param TokenStorageInterface $tokenStorage
      */
-    public function __construct(Environment $environment, TaskRepository $taskRepository, FormFactoryInterface $formFactory, EntityManagerInterface $entityManager, UrlGeneratorInterface $urlGenerator, FlashBagInterface $flashBag, TokenStorageInterface $tokenStorage)
+    public function __construct(Environment $environment, TaskRepository $taskRepository, FormFactoryInterface $formFactory, EntityManagerInterface $entityManager, UrlGeneratorInterface $urlGenerator, AuthorizationCheckerInterface $authorizationChecker, FlashBagInterface $flashBag, TokenStorageInterface $tokenStorage)
     {
         $this->environment = $environment;
         $this->taskRepository = $taskRepository;
         $this->formFactory = $formFactory;
         $this->entityManager = $entityManager;
         $this->urlGenerator = $urlGenerator;
+        $this->authorizationChecker = $authorizationChecker;
         $this->flashBag = $flashBag;
         $this->tokenStorage = $tokenStorage;
     }
 
     /**
      * @Route("/tasks", name="task_list")
-     * @IsGranted("IS_AUTHENTICATED_FULLY")
      */
     public function listAction()
     {
-        $tasks = $this->taskRepository->findAll();
+        /** @var User $user */
+        $user =  $this->tokenStorage->getToken()->getUser();
+
+        if ($this->authorizationChecker->isGranted('ROLE_ADMIN')) {
+            $tasks = $this->taskRepository->getAdminTasks($user);
+        } else {
+            $tasks = $this->taskRepository->getUserTasks($user);
+        }
 
         return new Response($this->environment->render(
             'task/list.html.twig',
@@ -93,6 +106,8 @@ class TaskController extends Controller
         if ($form->isValid() && $form->isSubmitted()) {
 
             $task->setUser($user);
+            $task->setIsAnonymous(false);
+
             $this->entityManager->persist($task);
             $this->entityManager->flush();
 
@@ -116,10 +131,9 @@ class TaskController extends Controller
     public function editAction(Task $task, Request $request)
     {
         $form = $this->formFactory->create(TaskType::class, $task);
-
         $form->handleRequest($request);
 
-        if ($form->isValid() && $form->isSubmitted()) {
+        if ($this->checkRights($task) && $form->isValid() && $form->isSubmitted()) {
             $this->entityManager->flush();
 
             $this->flashBag->add('success', 'La tâche a bien été modifiée.');
@@ -142,10 +156,12 @@ class TaskController extends Controller
      */
     public function toggleTaskAction(Task $task)
     {
-        $task->toggle(!$task->isDone());
-        $this->entityManager->flush();
+        if($this->checkRights($task)) {
+            $task->toggle(!$task->isDone());
+            $this->entityManager->flush();
 
-        $this->flashBag->add('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+            $this->flashBag->add('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+        }
 
         return new RedirectResponse($this->urlGenerator->generate('task_list'));
     }
@@ -156,11 +172,36 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
-        $this->entityManager->remove($task);
-        $this->entityManager->flush();
+        if($this->checkRights($task)) {
+            $this->entityManager->remove($task);
+            $this->entityManager->flush();
 
-        $this->flashBag->add('success', 'La tâche a bien été supprimée.');
+            $this->flashBag->add('success', 'La tâche a bien été supprimée.');
+        }
 
         return new RedirectResponse($this->urlGenerator->generate('task_list'));
+    }
+
+    /**
+     * Grant Admin Rights to Anonymous Tasks
+     * Grant Users/Admins Rights to their Tasks
+     * @param Task $task
+     * @return bool
+     */
+    private function checkRights(Task $task): bool
+    {
+        if ($task->getIsAnonymous()) {
+            // Task Has No User (isAnonymous), so Users cannot But Admin Can (return true)
+            if($this->authorizationChecker->isGranted('ROLE_ADMIN')) {
+                return true;
+            }
+        }
+
+        // tasks are not anonymous, check if the User of the task match the User connected
+        if ($task->getUser() == $this->tokenStorage->getToken()->getUser()) {
+            return true;
+        }
+
+        return false;
     }
 }
